Reorganize Model directory and refactor files that use it
